fix bug

+++ src/java/mg/itu/utilproject/modele/BaseModele.java
@@ -1,8 +1,4 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+
 package mg.itu.utilproject.modele;
 
 /**
@@ -10,13 +6,13 @@ package mg.itu.utilproject.modele;
  * @author davra
  */
 public class BaseModele {
-    protected long id;
+    protected int id;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 }
